COM-2260 remove moment, create specific class folder
