+++ package.json
@@ -73,7 +73,6 @@
     "expo-updates": "~0.8.0",
     "luxon": "^2.0.1",
     "mime": "^2.5.2",
-    "moment": "^2.29.1",
     "qs": "^6.10.1",
     "react": "16.13.1",
     "react-dom": "16.13.1",
